admitting defeat and fixing broken checks
