+++ ecs_model_deployer/src/lib/ecs-schema.ts
@@ -1,4 +1,3 @@
-/// <reference types="vite/client" />
 /**
   Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 
@@ -19,8 +18,12 @@
 import * as cdk from 'aws-cdk-lib';
 import { AmiHardwareType } from 'aws-cdk-lib/aws-ecs';
 import { z } from 'zod';
+import path from 'path';
+import fs from 'fs';
 
-import VERSION from '../VERSION.txt?raw';
+const HERE: string = path.resolve(__dirname);
+const VERSION_PATH: string = path.resolve(HERE, '..', 'VERSION');
+export const VERSION: string = fs.readFileSync(VERSION_PATH, 'utf8').trim();
 
 const REMOVAL_POLICIES: Record<string, cdk.RemovalPolicy> = {
     destroy: cdk.RemovalPolicy.DESTROY,
