+++ vector_store_deployer/src/lib/index.ts
@@ -16,7 +16,8 @@
 import { App } from 'aws-cdk-lib/core';
 import { OpenSearchVectorStoreStack } from './opensearch';
 import { PGVectorStoreStack } from './pgvector';
-import { PartialConfigSchema, RagRepositoryConfigSchema, RagRepositoryType } from '../../../lib/configSchema';
+import { RagRepositoryConfigSchema, RagRepositoryType } from '../../../lib/configSchema';
+import { PartialConfigSchema } from '../../../lib/schema';
 
 const app = new App();
