+++ vector_store_deployer/src/lib/opensearch.ts
@@ -16,7 +16,8 @@
 import { RemovalPolicy, StackProps } from 'aws-cdk-lib';
 import { Domain, EngineVersion, IDomain } from 'aws-cdk-lib/aws-opensearchservice';
 import { Construct } from 'constructs';
-import { PartialConfig, RagRepositoryConfig, RagRepositoryType } from '../../../lib/configSchema';
+import { RagRepositoryConfig, RagRepositoryType } from '../../../lib/configSchema';
+import { PartialConfig } from '../../../lib/schema';
 import { SecurityGroup, Subnet, SubnetSelection, Vpc } from 'aws-cdk-lib/aws-ec2';
 import { AnyPrincipal, CfnServiceLinkedRole, Effect, PolicyStatement, Role } from 'aws-cdk-lib/aws-iam';
 import { StringParameter } from 'aws-cdk-lib/aws-ssm';
