+++ vector_store_deployer/src/lib/pipeline-stack.ts
@@ -14,9 +14,9 @@
   limitations under the License.
 */
 import { Stack } from 'aws-cdk-lib';
-import * as z from 'zod';
 import { Construct } from 'constructs';
-import { PartialConfig, PipelineConfig, RagRepositoryConfig } from '../../../lib/configSchema';
+import { PipelineConfig, RagRepositoryConfig } from '../../../lib/configSchema';
+import { PartialConfig } from '../../../lib/schema';
 import { EventField, EventPattern, Rule, RuleTargetInput, Schedule } from 'aws-cdk-lib/aws-events';
 import { SfnStateMachine } from 'aws-cdk-lib/aws-events-targets';
 import { IStateMachine, StateMachine } from 'aws-cdk-lib/aws-stepfunctions';
